--- plugins/source/aws/resources/services/iam/accounts_fetch.go
@@ -1,1 +1,0 @@
-package iam
