--- plugins/source/aws/resources/services/iam/saml_identity_providers_fetch.go
@@ -1,1 +1,0 @@
-package iam
