--- plugins/source/aws/resources/services/iam/signing_certificates_fetch.go
@@ -1,1 +1,0 @@
-package iam
