+++ plugins/source/aws/resources/services/iam/user_attached_policies.go
@@ -1,6 +1,9 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -39,4 +42,21 @@ func userAttachedPolicies() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchIamUserAttachedPolicies(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	p := parent.Item.(*types.User)
+	svc := meta.(*client.Client).Services().Iam
+	config := iam.ListAttachedUserPoliciesInput{
+		UserName: p.UserName,
+	}
+	paginator := iam.NewListAttachedUserPoliciesPaginator(svc, &config)
+	for paginator.HasMorePages() {
+		output, err := paginator.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- output.AttachedPolicies
+	}
+	return nil
 }
